+++ NEWS.md
@@ -1,5 +1,5 @@
-  * Added an optional path for the X-Frame-Options header retrieval
-  * Disallowing ```--require-hsts``` and ```--require-x-frame-options``` if no HTTPS is used
+* 2022-09-09 Version 2.43.0
+  * Rewrote the HTTP security header checks
   * Fixed a bug when parsing certificates without purpose
 * 2022-09-02 Version 2.42.0
   * Disable nmap checks if a proxy is specified